14:20

+++ src/primitives/Util.java
@@ -18,12 +18,12 @@ public abstract class Util {
     // 1 bit sign, 11 bits exponent, 53 bits (52 stored) normalized mantissa
     // the number is m+2^e where 1<=m<2
     // NB: exponent is stored "normalized" (i.e. always positive by adding 1023)
-    private static int getExp(double num) {
+    private static double getExp(double num) {
         // 1. doubleToRawLongBits: "convert" the stored number to set of bits
         // 2. Shift all 52 bits to the right (removing mantissa)
         // 3. Zero the sign of number bit by mask 0x7FF
         // 4. "De-normalize" the exponent by subtracting 1023
-        return (int)((Double.doubleToRawLongBits(num) >> 52) & 0x7FFL) - 1023;
+        return (double)((Double.doubleToRawLongBits(num) >> 52) & 0x7FFL) - 1023;
     }
 
     /**
